Rewrite status as module
