+++ Gemfile.modules
@@ -38,6 +38,7 @@ group :opf_plugins do
 	gem 'openproject-github_integration',        path: 'modules/github_integration'
 	gem 'openproject-ldap_groups',               path: 'modules/ldap_groups'
 	gem 'openproject-recaptcha',                 path: 'modules/recaptcha'
+	gem 'openproject-job_status',                path: 'modules/job_status'
 
 	gem 'grids',                                 path: 'modules/grids'
 	gem 'my_page',                               path: 'modules/my_page'
