+++ frontend/src/app/modules/job-status/display-job-page/display-job-page.component.html
@@ -1,0 +1,1 @@
+Hello {{jobId}}
