+++ modules/job_status/db/migrate/20200610124259_extend_job_status.rb
