+++ modules/job_status/openproject-job_status.gemspec
@@ -1,0 +1,11 @@
+# encoding: UTF-8
+
+Gem::Specification.new do |s|
+  s.name        = 'openproject-job_status'
+  s.version     = '1.0.0'
+  s.authors     = 'OpenProject GmbH'
+  s.email       = 'info@openproject.com'
+  s.summary     = 'OpenProject Job status'
+  s.description = 'Listing and status of background jobs'
+  s.license     = 'GPLv3'
+end
